add switchtoggle element and refactor element boilerplate
