+++ Elements/ButtonToggleElement.swift
@@ -7,44 +7,11 @@
 
 import Foundation
 
-@IBDesignable class ButtonToggleElement: UIView {
+class ButtonToggleElement: ElementView {
     
-    @IBInspectable var contentView: UIView!
+    @IBOutlet var label: UILabel!
     
-    @IBInspectable @IBOutlet var label: UILabel!
-    
-    @IBInspectable @IBOutlet var button: UIButton!
-    
-    // for using ButtonToggleElement in code
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-        loadViewFromNIB()
-    }
-    
-    // for using ButtonToggleElement in IB
-    required init?(coder aDecoder: NSCoder) {
-        super.init(coder: aDecoder)
-        loadViewFromNIB()
-    }
-    
-    func loadViewFromNIB() {
-        let bundle = Bundle(for: type(of: self))
-        let nib = UINib(nibName: String(describing: type(of: self)), bundle: bundle)
-        let view = nib.instantiate(withOwner: self, options: nil).first as! UIView
-        
-        view.isUserInteractionEnabled = true
-        
-        view.frame = self.bounds
-        self.addSubview(view)
-        contentView = view
-    }
-    
-    override func prepareForInterfaceBuilder() {
-        super.prepareForInterfaceBuilder()
-        loadViewFromNIB()
-        contentView?.prepareForInterfaceBuilder()
-    }
-    
+    @IBOutlet var button: UIButton!
     
     @IBAction func buttonPressed(_ sender: UIButton) {
         print("button pressed")
