+++ Elements/GroupElement.swift
@@ -7,41 +7,10 @@
 
 import Foundation
 
-@IBDesignable class GroupElement: UIView {
+class GroupElement: ElementView {
     
-    @IBInspectable var contentView: UIView!
+    @IBOutlet var label: UILabel!
     
-    @IBInspectable @IBOutlet var label: UILabel!
-    
-    @IBInspectable @IBOutlet var stackView: UIStackView!
-    
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-        loadViewFromNIB()
-    }
-    
-    required init?(coder aDecoder: NSCoder) {
-        super.init(coder: aDecoder)
-        loadViewFromNIB()
-    }
-    
-    func loadViewFromNIB() {
-        let bundle = Bundle(for: type(of: self))
-        let nib = UINib(nibName: String(describing: type(of: self)), bundle: bundle)
-        let view = nib.instantiate(withOwner: self, options: nil).first as! UIView
-        
-        stackView.isUserInteractionEnabled = true
-        view.isUserInteractionEnabled = true
-        
-        view.frame = self.bounds
-        self.addSubview(view)
-        contentView = view
-    }
-    
-    override func prepareForInterfaceBuilder() {
-        super.prepareForInterfaceBuilder()
-        loadViewFromNIB()
-        contentView?.prepareForInterfaceBuilder()
-    }
+    @IBOutlet var stackView: UIStackView!
     
 }
