+++ Elements/GroupElement.xib
@@ -10,14 +10,14 @@
         <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
     </dependencies>
     <objects>
-        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner" customClass="GroupElement" customModule="DeviceDisplay" customModuleProvider="target">
+        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner" customClass="GroupElement" customModule="DHAP" customModuleProvider="target">
             <connections>
                 <outlet property="label" destination="fgJ-ZQ-0jK" id="AGl-xf-s2N"/>
                 <outlet property="stackView" destination="gT9-FO-zrm" id="ygt-Vj-S3S"/>
             </connections>
         </placeholder>
         <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
-        <view userInteractionEnabled="NO" contentMode="scaleToFill" id="iN0-l3-epB">
+        <view contentMode="scaleToFill" id="iN0-l3-epB">
             <rect key="frame" x="0.0" y="0.0" width="340" height="237"/>
             <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
             <subviews>
@@ -27,7 +27,7 @@
                     <nil key="textColor"/>
                     <nil key="highlightedColor"/>
                 </label>
-                <stackView opaque="NO" userInteractionEnabled="NO" contentMode="scaleToFill" axis="vertical" distribution="fillEqually" spacing="10" translatesAutoresizingMaskIntoConstraints="NO" id="gT9-FO-zrm">
+                <stackView opaque="NO" contentMode="scaleToFill" axis="vertical" distribution="fillEqually" spacing="10" translatesAutoresizingMaskIntoConstraints="NO" id="gT9-FO-zrm">
                     <rect key="frame" x="0.0" y="29" width="340" height="174"/>
                     <color key="backgroundColor" red="0.6776154902" green="0.78253807470000003" blue="0.90961211919999996" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
                 </stackView>
